Create PassengerApp
Reads in an xml file of flight passenger names and information and groups passengers by what class they are flying in. Creates a GUI where the user can select a class from set of check boxes, and a list of passengers from that class will appear in a jscrollpane. The user can then click on a passenger and view his/her information. User can switch from class to class and passenger to passenger at any time.